Add restricted achievements
The ability to add achievements that only admins
or specially designated users can grant, and only
admins can add. The idea is that this can be used
to create more structured achievements (e.g., for
tasks that a group wants as many members as
possible to complete could grant the achievements
for completion as the tasks are completed to track
who has done what). For restricted achievements,
add a third field to the achievement in the
achievements file, so the format is
name : description : users
where users is a list of users that can grant the
achievement (admins can grant any achievement).
If there is a second field seperator but no users,
then only admins can grant the achievement.