Change how data stored
Massive change to how things work. Previously,
everything was done in files, everything required
that the file be opened to read then closed with
everything being in the proper format and all of
the information being parsed on the fly. It hurt,
but then I remembered that computers have memory
to store data and python is an object-oriented
language, and I could use that to make it much
easier to work with. There's been very little
change in how Achievebot works under the hood,
but the big changes are in the format of the files
used to store data. In userfile, the format for a
user is now of the form
user -> achname1;achname2;achname3
and in achievefile, the format is
name : desc : restricted : perms
where the fields are the name of the achievement,
the description of the achievement, if the
achievement is restricted ('True' or 'False'),
and who has permissions to grant the achievement.
All of these fields can be added in the same way
as before, and the first two fields are required
in every achievement definition as before with
admins being able to make a restricted achievement
and add perms as before. All of this should
(hopefully) make things work a little better.